+++ node/middlewares/events/index.ts
@@ -1,5 +1,5 @@
-import { compareOrder } from './compareOrder'
 import { eventsErrorHandler } from './eventsErrorHandler'
+import { modifyGlovoOrder } from './modifyGlovoOrder'
 import { updateGlovoOrderStatus } from './updateGlovoOrderStatus'
 import { validateEventSettings } from './validateEventSettings'
 
@@ -7,7 +7,7 @@ export const events = {
   updateGlovoOrder: [
     eventsErrorHandler,
     validateEventSettings,
-    compareOrder,
+    modifyGlovoOrder,
     updateGlovoOrderStatus,
   ],
 }
